fix: addCoount (증가할 좋아요 수) 데이터 타입 long -> integer로 변경, request body 형식 변경

+++ src/main/java/likelion/festival/service/PubService.java
@@ -29,7 +29,7 @@ public class PubService {
     }
 
     @Transactional
-    public void addPubLike(Long pubId, Long addCount) {
+    public void addPubLike(Long pubId, Integer addCount) {
         Pub pub = pubRepository.findById(pubId)
                 .orElseThrow(() -> new EntityNotFoundException("요청한 id를 가진 주점을 찾을 수 없습니다: " + pubId));
         pub.addLikeCount(addCount);
